+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.7.0
 
 *   Add literals and typedefs for km/h (`_km_p_h`) m/s (`m_p_s`)
-*   Add literals for astronomical units (AU) and lightyears (ly)
+*   Add literals for astronomical units of length (AU), lightyears (ly) and parsecs (pc)
 
 ## 1.6.1